+++ src/main/java/hanium/modic/backend/web/ai/dto/response/GetTicketInformationResponse.java
@@ -6,13 +6,13 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 public record GetTicketInformationResponse(
 	@Schema(description = "남은 티켓 수")
-	Integer ticketCount,
+	Long ticketCount,
 
 	@Schema(description = "다음 리셋까지 시간")
 	LocalDateTime nextReset
 ) {
 
-	public static GetTicketInformationResponse of(int ticketCount, LocalDateTime nextReset) {
+	public static GetTicketInformationResponse of(long ticketCount, LocalDateTime nextReset) {
 		return new GetTicketInformationResponse(
 			ticketCount,
 			nextReset
